Define mock modules at runtime
Before, we were using defmodule to define mocks. This presents an issue when defining
mocks within a function called by multiple tests running in parallel: the same mock is defined
multiple times with the same name because the name is generated at compile time.

Now, we use Module.create/3 to accomplish the same thing at runtime so we can dynamically define
multiple mocks with the same function.